backend/requirements.txt
fastapi
uvicorn[standard]
google-adk
pydantic
GitPython
prometheus_client
python-json-logger
black
isort
python-dotenv
requests
ruff